requirements.txt
numpy
opencv-python-headless
pillow
albumentations
requests
jiwer
tqdm
